moved SegmentationTrainingTeiParser
