+++ sciencebeam_parser/models/segmentation/training_data.py
@@ -1,16 +1,23 @@
 import logging
-from typing import Iterable, List
+from typing import Iterable, List, Tuple, Union
 
+from lxml import etree
 from lxml.builder import ElementMaker
 
 from sciencebeam_parser.utils.xml_writer import XmlTreeWriter
+from sciencebeam_parser.utils.xml import get_text_content
+from sciencebeam_parser.utils.tokenizer import get_tokenized_tokens
 from sciencebeam_parser.document.layout_document import (
     LayoutLine,
     LayoutToken,
     join_layout_tokens
 )
 from sciencebeam_parser.models.data import LayoutModelData
-from sciencebeam_parser.models.model import get_split_prefix_label
+from sciencebeam_parser.models.model import (
+    NEW_DOCUMENT_MARKER,
+    NewDocumentMarker,
+    get_split_prefix_label
+)
 from sciencebeam_parser.models.training_data import (
     AbstractTeiTrainingDataGenerator,
     get_model_data_label
@@ -129,3 +136,50 @@ class SegmentationTeiTrainingDataGenerator(AbstractTeiTrainingDataGenerator):
                 pending_whitespace = '\n'
         xml_writer.require_path(default_path)
         xml_writer.append_text(pending_whitespace)
+
+
+def iter_tag_result_for_flat_tag_result(
+    flat_tag_result_iterable: Iterable[Union[Tuple[str, str], NewDocumentMarker]]
+) -> Iterable[List[Tuple[str, str]]]:
+    doc_tag_result: List[Tuple[str, str]] = []
+    for token_tag_result in flat_tag_result_iterable:
+        if isinstance(token_tag_result, NewDocumentMarker):
+            yield doc_tag_result
+            doc_tag_result = []
+            continue
+        doc_tag_result.append(token_tag_result)
+
+
+def get_tag_result_for_flat_tag_result(
+    flat_tag_result_iterable: Iterable[Union[Tuple[str, str], NewDocumentMarker]]
+) -> List[List[Tuple[str, str]]]:
+    return list(iter_tag_result_for_flat_tag_result(flat_tag_result_iterable))
+
+
+class SegmentationTrainingTeiParser:
+    def iter_parse_training_tei_to_flat_tag_result(
+        self,
+        tei_root: etree.ElementBase
+    ) -> Iterable[Union[Tuple[str, str], NewDocumentMarker]]:
+        for text_node in tei_root.xpath('./text'):
+            if text_node.text:
+                for token_text in get_tokenized_tokens(text_node.text):
+                    yield token_text, 'O'
+            for child_node in text_node:
+                label = '<' + child_node.tag + '>'
+                for token_index, token_text in enumerate(
+                    get_tokenized_tokens(get_text_content(child_node))
+                ):
+                    prefix = 'B-' if token_index == 0 else 'I-'
+                    yield token_text, prefix + label
+            yield NEW_DOCUMENT_MARKER
+
+    def parse_training_tei_to_tag_result(
+        self,
+        tei_root: etree.ElementBase
+    ):
+        return get_tag_result_for_flat_tag_result(
+            self.iter_parse_training_tei_to_flat_tag_result(
+                tei_root
+            )
+        )
