+++ sciencebeam_parser/training/cli/generate_delft_data.py
@@ -2,7 +2,7 @@ import argparse
 import logging
 from glob import glob
 from pathlib import Path
-from typing import Iterable, List, Optional, Tuple, Union
+from typing import List, Optional, Tuple
 
 from lxml import etree
 
@@ -12,60 +12,13 @@ from sciencebeam_trainer_delft.sequence_labelling.tag_formatter import (
     TagOutputFormats,
     format_tag_result
 )
-from sciencebeam_parser.models.model import NEW_DOCUMENT_MARKER, NewDocumentMarker
 
-from sciencebeam_parser.utils.xml import get_text_content
-from sciencebeam_parser.utils.tokenizer import get_tokenized_tokens
+from sciencebeam_parser.models.segmentation.training_data import (
+    SegmentationTrainingTeiParser
+)
 
 
 LOGGER = logging.getLogger(__name__)
-
-
-def iter_tag_result_for_flat_tag_result(
-    flat_tag_result_iterable: Iterable[Union[Tuple[str, str], NewDocumentMarker]]
-) -> Iterable[List[Tuple[str, str]]]:
-    doc_tag_result: List[Tuple[str, str]] = []
-    for token_tag_result in flat_tag_result_iterable:
-        if isinstance(token_tag_result, NewDocumentMarker):
-            yield doc_tag_result
-            doc_tag_result = []
-            continue
-        doc_tag_result.append(token_tag_result)
-
-
-def get_tag_result_for_flat_tag_result(
-    flat_tag_result_iterable: Iterable[Union[Tuple[str, str], NewDocumentMarker]]
-) -> List[List[Tuple[str, str]]]:
-    return list(iter_tag_result_for_flat_tag_result(flat_tag_result_iterable))
-
-
-class SegmentationTrainingTeiParser:
-    def iter_parse_training_tei_to_flat_tag_result(
-        self,
-        tei_root: etree.ElementBase
-    ) -> Iterable[Union[Tuple[str, str], NewDocumentMarker]]:
-        for text_node in tei_root.xpath('./text'):
-            if text_node.text:
-                for token_text in get_tokenized_tokens(text_node.text):
-                    yield token_text, 'O'
-            for child_node in text_node:
-                label = '<' + child_node.tag + '>'
-                for token_index, token_text in enumerate(
-                    get_tokenized_tokens(get_text_content(child_node))
-                ):
-                    prefix = 'B-' if token_index == 0 else 'I-'
-                    yield token_text, prefix + label
-            yield NEW_DOCUMENT_MARKER
-
-    def parse_training_tei_to_tag_result(
-        self,
-        tei_root: etree.ElementBase
-    ):
-        return get_tag_result_for_flat_tag_result(
-            self.iter_parse_training_tei_to_flat_tag_result(
-                tei_root
-            )
-        )
 
 
 def parse_segmentation_training_tei_to_tag_result(
